@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 8
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.dynamite.heaterrc"
-        minSdkVersion 8 //14
-        targetSdkVersion 8 //22, 28
-        versionCode 16
-        versionName "1.6"
+        minSdkVersion 9 //14
+        targetSdkVersion 34 //22, 28
+        versionCode 22
+        versionName "2.2"
     }
 
     buildTypes {
