@@ -1,5 +1,5 @@
 package com.dynamite.heaterrc;
-/* 
+/*
 DisplayNotifications.java
 
 Copyright (C) 2015  dynamitetuning
@@ -18,33 +18,23 @@ You should have received a copy of the GNU General Public License
 along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
-import android.app.Activity;
-import android.app.Notification;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
-import android.content.SharedPreferences;
+
 import android.os.Bundle;
-import android.telephony.SmsManager;
-import android.widget.Toast;
- 
+
 public class DisplayNotifications extends commonActivity {
     /** Called when the activity is first created. */
 	// private static final String DEBUG_TAG = "DisplayNotifications";
     @Override
     public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);     
-        
+        super.onCreate(savedInstanceState);
+        /*
         final SharedPreferences settings = getSharedPreferences(PREFS_NAME, MODE_PRIVATE);
         final SharedPreferences.Editor SPeditor = settings.edit();
-        
+
         try {
 	        // Register broadcast receivers for SMS sent and delivered intents
 	        registerReceiver(new BroadcastReceiver() {
-	            
+
 	        	@Override
 	            public void onReceive(Context context, Intent intent) {
 	                String message = null;
@@ -67,7 +57,7 @@ public class DisplayNotifications extends commonActivity {
 	                    message = "Error: Radio off.";
 	                    break;
 	                default:
-	                    message = "Unknown Error.";	
+	                    message = "Unknown Error.";
 	                }
 	                // Log.d(DEBUG_TAG, "Broadcast received="+message);
 	                String temp = settings.getString(getString(R.string.sp_lastAlarm), "-");
@@ -84,8 +74,8 @@ public class DisplayNotifications extends commonActivity {
         } catch (Exception ex) {
         	ex.printStackTrace();
         }
-        
-        //---get the notification ID for the notification; 
+
+        //---get the notification ID for the notification;
         // passed in by the MainActivity---
 		int notifID = 0;
 		try {
@@ -93,34 +83,35 @@ public class DisplayNotifications extends commonActivity {
 		} catch (NullPointerException npe){
         	npe.printStackTrace();
 		}
-        //---PendingIntent to launch activity if the user selects 
+        //---PendingIntent to launch activity if the user selects
         // the notification---
         Intent i = new Intent("com.dynamite.heaterrc.AlarmDetails");
-        i.putExtra("NotifID", notifID);  
- 
-        PendingIntent detailsIntent = 
-            PendingIntent.getActivity(this, 0, i, 0);
- 
+        i.putExtra("NotifID", notifID);
+
+        PendingIntent detailsIntent =
+            PendingIntent.getActivity(this, 0, i, PendingIntent.FLAG_IMMUTABLE);
+
         NotificationManager nm = (NotificationManager)
             getSystemService(NOTIFICATION_SERVICE);
         Notification notif = new Notification(
-            R.drawable.ic_launcher, 
+            R.drawable.ic_launcher,
             getString(R.string.dn_notificationTitle),
             System.currentTimeMillis());
 
         CharSequence from = getString(R.string.dn_notificationTitle);
-        CharSequence message = getString(R.string.dn_notificationText);        
+        CharSequence message = getString(R.string.dn_notificationText);
         notif.setLatestEventInfo(this, from, message, detailsIntent);
- 
+
         //---100ms delay, vibrate for 250ms, pause for 100 ms and
         // then vibrate for 500ms---
-        // notif.vibrate = new long[] { 100, 250, 100, 500};        
+        // notif.vibrate = new long[] { 100, 250, 100, 500};
+        assert nm != null;
         nm.notify(notifID, notif);
         //---destroy the activity---
-        
-        finish();
+
+        finish(); */
     }
-    
+
     public void onResume(Bundle savedInstaceState){
     	// nothing to do
     }
